refactor: rename tasks to handlers
