+++ firestore-incremental-capture/functions/src/handlers/on_backup_restore_handler.ts
