+++ firestore-incremental-capture/functions/src/handlers/restore_done_trigger_handler.ts
