+++ firestore-incremental-capture/functions/src/handlers/run_initial_setup_handler.ts
