+++ firestore-incremental-capture/functions/src/handlers/sync_data_handler.ts
