+++ firestore-incremental-capture/functions/src/handlers/trigger_restoration_job_handler.ts
